+++ src/module.test.ts
@@ -79,6 +79,7 @@ describe('TestPlatform', () => {
   });
 
   it('should return an instance of the platform', async () => {
+    matterbridge.matterbridgeVersion = '3.5.1';
     dynamicPlatform = initializePlugin(matterbridge, log, config);
     expect(dynamicPlatform).toBeInstanceOf(ExampleMatterbridgeDynamicPlatform);
     await dynamicPlatform.onShutdown();
@@ -87,9 +88,9 @@ describe('TestPlatform', () => {
   it('should throw error in load when version is not valid', () => {
     matterbridge.matterbridgeVersion = '1.5.0';
     expect(() => new ExampleMatterbridgeDynamicPlatform(matterbridge, log, config)).toThrow(
-      'This plugin requires Matterbridge version >= "3.4.0". Please update Matterbridge from 1.5.0 to the latest version in the frontend.',
+      'This plugin requires Matterbridge version >= "3.5.1". Please update Matterbridge from 1.5.0 to the latest version in the frontend.',
     );
-    matterbridge.matterbridgeVersion = '3.4.0';
+    matterbridge.matterbridgeVersion = '3.5.1';
   });
 
   it('should initialize platform with config name and set the default config', () => {