+++ CHANGELOG.md
@@ -32,6 +32,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
 
+
 ## [2.0.4] - 2026-01-20
 
 ### Added