+++ CHANGELOG.md
@@ -36,6 +36,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 - [package]: Updated dependencies.
 - [package]: Bumped package to automator v.3.0.2.
+- [package]: Required matterbridge 3.5.1.
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>