+++ CHANGELOG.md
@@ -27,7 +27,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Added
 
-- [thermostat]: Added a thermostat auto mode with presets (Comfort and Away modes) including 3 sub endpoints with flowMeasurement, temperature and humidity sensors.
+- [thermostat]: Added a thermostat auto mode with presets (Comfort and Away modes) including 3 sub endpoints with flowMeasurement, temperature and humidity sensors. Thanks Ludovic BOUÉ (<https://github.com/Luligu/matterbridge-example-dynamic-platform/pull/42>).
 - [presets]: Added support for Matter Thermostat Presets feature with preset activation and setpoint management.
 - [Energy]: Added an outlet (plug) with onOff cluster, energy measurements and power measurements
 - [EnergyApparent]: Added an outlet (plug) with onOff cluster, apparent energy measurements and power measurements