+++ CHANGELOG.md
@@ -23,7 +23,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
-## [2.0.5] - 2026-01-24
+## [2.0.5] - 2026-01-26
 
 ### Added